requirements.txt
streamlit
easyocr
opencv-python-headless
Pillow
numpy
language-tool-python
torch==1.13.1
torchvision==0.14.1

